+++ fileutils/fileutils.go
@@ -1,4 +1,4 @@
-package utils
+package fileutils
 
 import (
 	"os"
